soon v1

+++ src/services/CompanyService.ts
@@ -1,17 +1,17 @@
 /* eslint-disable import/no-dynamic-require */
 import { NotFoundError } from '../handlers/ApiError';
 import { dbDir } from '../config';
-import { ICompany, ICompanyInvite } from '../databases/interfaces';
+import { ICompany, IInvite } from '../databases/interfaces';
 
 const dbPath = `../databases/${dbDir}`;
 const { default: CompanyApi } = require(`${dbPath}/api/CompanyApi`);
-const { default: CompanyInviteApi } = require(`${dbPath}/api/CompanyInviteApi`);
+const { default: InviteApi } = require(`${dbPath}/api/InviteApi`);
 
 export default class CompanyService {
   /* CRUD */
-  public static async getById(id: any): Promise<ICompany | undefined> {
+  public static async getById(id: any): Promise<ICompany | null> {
     const company: ICompany | undefined = await CompanyApi.getById(id);
-    if (!company) throw new NotFoundError('Company not found!');
+    if (!company) throw new NotFoundError(`Company "${id}" is not found!`);
     return company;
   }
 
@@ -19,28 +19,44 @@ export default class CompanyService {
     return CompanyApi.getAll();
   }
 
-  public static async createOne(companyData: ICompany): Promise<ICompany> {
-    return CompanyApi.createOne(companyData);
+  public static async createOne(companyData: ICompany, creatorId: any): Promise<ICompany> {
+    const company: ICompany = { ...companyData, employees: [{ user: creatorId, isAdmin: true }] };
+    const insertedCompany = await CompanyApi.createOne(company);
+    return { id: insertedCompany.id, ...company };
   }
 
   public static async updateById(id: any, companyData: ICompany): Promise<ICompany | null> {
+    if (!await CompanyApi.isExists(id)) {
+      throw new NotFoundError(`Company "${id}" is not found!`);
+    }
     return CompanyApi.updateById(id, companyData);
   }
 
   public static async deleteById(id: any): Promise<ICompany> {
+    if (!await CompanyApi.isExists(id)) {
+      throw new NotFoundError(`Company "${id}" is not found!`);
+    }
     return CompanyApi.deleteById(id);
   }
   /* CRUD END */
 
-  public static async inviteUser(companyId: any, userId: any): Promise<ICompanyInvite> {
-    return CompanyInviteApi.createOne({ company: companyId, user: userId });
+  public static async isUserInCompany(companyId: any, userId: any): Promise<boolean> {
+    return CompanyApi.isUserInCompany(companyId, userId);
   }
 
-  public static async addUser(companyId: any, userId: any): Promise<ICompany | null> {
-    return CompanyApi.addUser(companyId, userId);
+  public static async isUserCompanyAdmin(companyId: any, userId: any): Promise<boolean> {
+    return CompanyApi.isUserCompanyAdmin(companyId, userId);
   }
 
-  public static async removeUser(companyId: any, userId: any): Promise<ICompany | null> {
-    return CompanyApi.removeUser(companyId, userId);
-  }
+  // public static async inviteUser(companyId: any, userId: any): Promise<IInvite> {
+  //   return InviteApi.createOne({ company: companyId, user: userId });
+  // }
+
+  // public static async addUser(companyId: any, userId: any): Promise<ICompany | null> {
+  //   return CompanyApi.addUser(companyId, userId);
+  // }
+
+  // public static async removeUser(companyId: any, userId: any): Promise<ICompany | null> {
+  //   return CompanyApi.removeUser(companyId, userId);
+  // }
 }
